Added all commands. Appears to be working.
