+++ input.c
@@ -563,984 +563,7 @@ int cmd_handle(char *inp, char **qmsg, fd_set *master, int *fdmax) // old state=
 		return(talk(cmd));
 	char *args=strchr(cmd, ' ');
 	if(args) *args++=0;
-	if(call_cmd(cmd,args,qmsg, fdmax)){
-		return -1;
-	}
-	else
-	{
-		return 0;
-	}
-	if(strcmp(cmd, "reconnect")==0)
-	{
-		if(bufs[cbuf].server)
-		{
-			if(!SERVER(cbuf).live)
-			{
-				char *newport;
-				if(args)
-				{
-					newport=args;
-				}
-				else if(SERVER(cbuf).autoent && SERVER(cbuf).autoent->portno)
-				{
-					newport=SERVER(cbuf).autoent->portno;
-				}
-				else
-				{
-					newport=portno;
-				}
-				char dstr[30+strlen(SERVER(cbuf).serverloc)+strlen(newport)];
-				sprintf(dstr, "Connecting to %s on port %s...", SERVER(cbuf).serverloc, newport);
-				#if ASYNCH_NL
-				nl_list *nl=irc_connect(SERVER(cbuf).serverloc, newport);
-				if(nl)
-				{
-					nl->reconn_b=bufs[cbuf].server;
-					add_to_buffer(bufs[cbuf].server, STA, QUIET, 0, false, dstr, "/server: ");
-					redraw_buffer();
-				}
-				else
-				{
-					add_to_buffer(bufs[cbuf].server, ERR, NORMAL, 0, false, "malloc failure (see status)", "/server: ");
-					redraw_buffer();
-				}
-				#else /* ASYNCH_NL */
-				int serverhandle=irc_connect(SERVER(cbuf).serverloc, newport, master, fdmax);
-				if(serverhandle)
-				{
-					int b=bufs[cbuf].server;
-					bufs[b].handle=serverhandle;
-					int b2;
-					for(b2=1;b2<nbufs;b2++)
-					{
-						if(bufs[b2].server==b)
-							bufs[b2].handle=serverhandle;
-					}
-					bufs[cbuf].conninpr=true;
-					free(bufs[cbuf].realsname);
-					bufs[cbuf].realsname=NULL;
-					add_to_buffer(cbuf, STA, QUIET, 0, false, dstr, "/server: ");
-				}
-				#endif /* ASYNCH_NL */
-			}
-			else
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Already connected to server", "/reconnect: ");
-			}
-		}
-		else
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must be run in the context of a server!", "/reconnect: ");
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "disconnect")==0)
-	{
-		int b=bufs[cbuf].server;
-		if(b>0)
-		{
-			if(bufs[b].handle)
-			{
-				if(bufs[b].live)
-				{
-					char quit[7+strlen(args?args:(qmsg&&*qmsg)?*qmsg:"quIRC quit")];
-					sprintf(quit, "QUIT %s", args?args:(qmsg&&*qmsg)?*qmsg:"quIRC quit");
-					irc_tx(bufs[b].handle, quit);
-				}
-				close(bufs[b].handle);
-				FD_CLR(bufs[b].handle, master);
-			}
-			int b2;
-			for(b2=1;b2<nbufs;b2++)
-			{
-				while((b2<nbufs) && (bufs[b2].type!=SERVER) && ((bufs[b2].server==b) || (bufs[b2].server==0)))
-				{
-					bufs[b2].live=false;
-					free_buffer(b2);
-				}
-			}
-			bufs[b].live=false;
-			free_buffer(b);
-			cbuf=0;
-			redraw_buffer();
-		}
-		else
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't disconnect (status)!", "/disconnect: ");
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "realsname")==0)
-	{
-		int b=bufs[cbuf].server;
-		if(b>0)
-		{
-			if(bufs[b].realsname)
-				add_to_buffer(cbuf, STA, NORMAL, 0, false, bufs[b].realsname, "/realsname: ");
-			else
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "unknown", "/realsname ");
-		}
-		else
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "(status) is not a server", "/realsname: ");
-		return(0);
-	}
-	if(strcmp(cmd, "join")==0)
-	{
-		if(!SERVER(cbuf).handle)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must be run in the context of a server!", "/join: ");
-		}
-		else if(!SERVER(cbuf).live)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Disconnected, can't send", "/join: ");
-		}
-		else if(args)
-		{
-			char *chan=strtok(args, " ");
-			if(!chan)
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify a channel!", "/join: ");
-			}
-			else
-			{
-				char *pass=strtok(NULL, ", ");
-				servlist *serv=SERVER(cbuf).autoent;
-				if(!serv)
-				{
-					serv=SERVER(cbuf).autoent=malloc(sizeof(servlist));
-					serv->name=NULL;
-					serv->portno=NULL;
-					serv->nick=NULL;
-					serv->pass=NULL;
-					serv->chans=NULL;
-					serv->next=NULL;
-					serv->igns=NULL;
-				}
-				if(pass)
-				{
-					chanlist *curr=malloc(sizeof(chanlist));
-					if(curr)
-					{
-						curr->name=strdup(chan);
-						curr->key=strdup(pass);
-						curr->next=serv->chans;
-						serv->chans=curr;
-					}
-				}
-				else
-				{
-					chanlist *curr=serv->chans;
-					while(curr)
-					{
-						if(irc_strcasecmp(curr->name, chan, SERVER(cbuf).casemapping)==0)
-						{
-							pass=curr->key;
-							break;
-						}
-						curr=curr->next;
-					}
-				}
-				if(!pass) pass="";
-				char joinmsg[8+strlen(chan)+strlen(pass)];
-				sprintf(joinmsg, "JOIN %s %s", chan, pass);
-				irc_tx(SERVER(cbuf).handle, joinmsg);
-			}
-		}
-		else
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify a channel!", "/join: ");
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "rejoin")==0)
-	{
-		if(bufs[cbuf].type==PRIVATE)
-		{
-			if(!(SERVER(cbuf).handle && SERVER(cbuf).live))
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Disconnected, can't send", "/rejoin: ");
-			}
-			else if(bufs[cbuf].live)
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Already in this channel", "/rejoin: ");
-			}
-			else
-			{
-				bufs[cbuf].live=true;
-			}
-		}
-		else if(bufs[cbuf].type!=CHANNEL)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "View is not a channel!", "/rejoin: ");
-		}
-		else if(!(SERVER(cbuf).handle && SERVER(cbuf).live))
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Disconnected, can't send", "/rejoin: ");
-		}
-		else if(bufs[cbuf].live)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Already in this channel", "/rejoin: ");
-		}
-		else
-		{
-			char *chan=bufs[cbuf].bname;
-			char *pass=args;
-			if(pass)
-				bufs[cbuf].lastkey=strdup(pass);
-			else
-				pass=bufs[cbuf].key;
-			if(!pass) pass="";
-			char joinmsg[8+strlen(chan)+strlen(pass)];
-			sprintf(joinmsg, "JOIN %s %s", chan, pass);
-			irc_tx(SERVER(cbuf).handle, joinmsg);
-			redraw_buffer();
-		}
-		return(0);
-	}
-	if((strcmp(cmd, "part")==0)||(strcmp(cmd, "leave")==0))
-	{
-		if(bufs[cbuf].type!=CHANNEL)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "This view is not a channel!", "/part: ");
-		}
-		else
-		{
-			if(LIVE(cbuf) && SERVER(cbuf).handle)
-			{
-				char partmsg[8+strlen(bufs[cbuf].bname)];
-				sprintf(partmsg, "PART %s", bufs[cbuf].bname);
-				irc_tx(SERVER(cbuf).handle, partmsg);
-				add_to_buffer(cbuf, PART, NORMAL, 0, true, "Leaving", "/part: ");
-			}
-			// when you try to /part a dead tab, interpret it as a /close
-			int parent=bufs[cbuf].server;
-			bufs[cbuf].live=false;
-			free_buffer(cbuf);
-			cbuf=parent;
-			redraw_buffer();
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "unaway")==0)
-	{
-		cmd="away";
-		args="-";
-	}
-	if(strcmp(cmd, "away")==0)
-	{
-		const char *am="Gone away, gone away, was it one of you took it away?";
-		if(args)
-			am=args;
-		if(SERVER(cbuf).handle)
-		{
-			if(LIVE(cbuf))
-			{
-				if(strcmp(am, "-"))
-				{
-					char nmsg[8+strlen(am)];
-					sprintf(nmsg, "AWAY :%s", am);
-					irc_tx(SERVER(cbuf).handle, nmsg);
-				}
-				else
-				{
-					irc_tx(SERVER(cbuf).handle, "AWAY"); // unmark away
-				}
-			}
-			else
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/away: ");
-			}
-		}
-		else
-		{
-			if(cbuf)
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/away: ");
-			else
-			{
-				int b;
-				for(b=0;b<nbufs;b++)
-				{
-					if(bufs[b].type==SERVER)
-					{
-						if(bufs[b].handle)
-						{
-							if(LIVE(b))
-							{
-								if(strcmp(am, "-"))
-								{
-									char nmsg[8+strlen(am)];
-									sprintf(nmsg, "AWAY :%s", am);
-									irc_tx(bufs[b].handle, nmsg);
-								}
-								else
-								{
-									irc_tx(bufs[b].handle, "AWAY"); // unmark away
-								}
-							}
-							else
-								add_to_buffer(b, ERR, NORMAL, 0, false, "Tab not live, can't send", "/away: ");
-						}
-						else
-							add_to_buffer(b, ERR, NORMAL, 0, false, "Tab not live, can't send", "/away: ");
-					}
-				}
-			}
-		}
-		return(0);
-	}
-	bool aalloc=false;
-	if(strcmp(cmd, "afk")==0)
-	{
-		const char *afm="afk";
-		if(args)
-			afm=strtok(args, " ");
-		const char *p=SERVER(cbuf).nick;
-		int n=strcspn(p, "|");
-		char *nargs=malloc(n+strlen(afm)+2);
-		if(nargs)
-		{
-			cmd="nick";
-			strncpy(nargs, p, n);
-			nargs[n]=0;
-			if(strcmp(afm, "-"))
-			{
-				strcat(nargs, "|");
-				strcat(nargs, afm);
-			}
-			args=nargs;
-			aalloc=true;
-		}
-	}
-	if(strcmp(cmd, "nick")==0)
-	{
-		if(args)
-		{
-			char *nn=strtok(args, " ");
-			if(SERVER(cbuf).handle)
-			{
-				if(LIVE(cbuf))
-				{
-					free(SERVER(cbuf).nick);
-					SERVER(cbuf).nick=strdup(nn);
-					char nmsg[8+strlen(SERVER(cbuf).nick)];
-					sprintf(nmsg, "NICK %s", SERVER(cbuf).nick);
-					irc_tx(SERVER(cbuf).handle, nmsg);
-					add_to_buffer(cbuf, STA, QUIET, 0, false, "Changing nick", "/nick: ");
-				}
-				else
-				{
-					add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/nick: ");
-				}
-			}
-			else
-			{
-				if(cbuf)
-					add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/nick: ");
-				else
-				{
-					free(bufs[0].nick);
-					bufs[0].nick=strdup(nn);
-					defnick=false;
-					add_to_buffer(cbuf, STA, QUIET, 0, false, "Default nick changed", "/nick: ");
-				}
-			}
-		}
-		else
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify a nickname!", "/nick: ");
-		}
-		if(aalloc) free(args);
-		return(0);
-	}
-	if(strcmp(cmd, "topic")==0)
-	{
-		if(args)
-		{
-			if(bufs[cbuf].type==CHANNEL)
-			{
-				if(SERVER(cbuf).handle)
-				{
-					if(LIVE(cbuf))
-					{
-						char tmsg[10+strlen(bufs[cbuf].bname)+strlen(args)];
-						sprintf(tmsg, "TOPIC %s :%s", bufs[cbuf].bname, args);
-						irc_tx(SERVER(cbuf).handle, tmsg);
-					}
-					else
-					{
-						add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/topic: ");
-					}
-				}
-				else
-				{
-					add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't send to channel - not connected!", "/topic: ");
-				}
-			}
-			else
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't set topic - view is not a channel!", "/topic: ");
-			}
-		}
-		else
-		{
-			if(bufs[cbuf].type==CHANNEL)
-			{
-				if(SERVER(cbuf).handle)
-				{
-					if(LIVE(cbuf))
-					{
-						char tmsg[8+strlen(bufs[cbuf].bname)];
-						sprintf(tmsg, "TOPIC %s", bufs[cbuf].bname);
-						irc_tx(SERVER(cbuf).handle, tmsg);
-					}
-					else
-					{
-						add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/topic: ");
-					}
-				}
-				else
-				{
-					add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't send to channel - not connected!", "/topic: ");
-				}
-			}
-			else
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't get topic - view is not a channel!", "/topic: ");
-			}
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "msg")==0)
-	{
-		if(!SERVER(cbuf).handle)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must be run in the context of a server!", "/msg: ");
-		}
-		else if(args)
-		{
-			bool no_tab=false;
-			char *dest=strtok(args, " ");
-			if(strcmp(dest, "-n")==0)
-			{
-				no_tab=true;
-				dest=strtok(NULL, " ");
-			}
-			if(!dest)
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify a recipient!", "/msg: ");
-				return(0);
-			}
-			char *text=strtok(NULL, "");
-			if(!no_tab)
-			{
-				int b2=makeptab(bufs[cbuf].server, dest);
-				cbuf=b2;
-				redraw_buffer();
-			}
-			if(text)
-			{
-				if(SERVER(cbuf).handle)
-				{
-					if(LIVE(cbuf))
-					{
-						char privmsg[12+strlen(dest)+strlen(text)];
-						sprintf(privmsg, "PRIVMSG %s :%s", dest, text);
-						irc_tx(SERVER(cbuf).handle, privmsg);
-						ctcp_strip(text, SERVER(cbuf).nick, cbuf, false, false, true, true);
-						if(no_tab)
-						{
-							add_to_buffer(cbuf, STA, QUIET, 0, false, "sent", "/msg -n: ");
-						}
-						else
-						{
-							while(text[strlen(text)-1]=='\n')
-								text[strlen(text)-1]=0; // stomp out trailing newlines, they break things
-							add_to_buffer(cbuf, MSG, NORMAL, 0, true, text, SERVER(cbuf).nick);
-						}
-					}
-					else
-					{
-						add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/msg: ");
-					}
-				}
-				else
-				{
-					add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't send to channel - not connected!", "/msg: ");
-				}
-			}
-		}
-		else
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify a recipient!", "/msg: ");
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "ping")==0)
-	{
-		if(!SERVER(cbuf).handle)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must be run in the context of a server!", "/ping: ");
-		}
-		else
-		{
-			const char *dest=NULL;
-			if(args)
-				dest=strtok(args, " ");
-			if(!dest&&bufs[cbuf].type==PRIVATE)
-				dest=bufs[cbuf].bname;
-			if(dest)
-			{
-				if(SERVER(cbuf).handle)
-				{
-					if(LIVE(cbuf))
-					{
-						struct timeval tv;
-						gettimeofday(&tv, NULL);
-						char privmsg[64+strlen(dest)];
-						snprintf(privmsg, 64+strlen(dest), "PRIVMSG %s :\001PING %u %u\001", dest, (unsigned int)tv.tv_sec, (unsigned int)tv.tv_usec);
-						irc_tx(SERVER(cbuf).handle, privmsg);
-					}
-					else
-					{
-						add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/ping: ");
-					}
-				}
-				else
-				{
-					add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't send - not connected!", "/ping: ");
-				}
-			}
-			else
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify a recipient!", "/ping: ");
-			}
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "amsg")==0)
-	{
-		if(!bufs[cbuf].server)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must be run in the context of a server!", "/amsg: ");
-		}
-		else if(args)
-		{
-			int b2;
-			for(b2=1;b2<nbufs;b2++)
-			{
-				if((bufs[b2].server==bufs[cbuf].server) && (bufs[b2].type==CHANNEL))
-				{
-					if(LIVE(b2))
-					{
-						char privmsg[12+strlen(bufs[b2].bname)+strlen(args)];
-						sprintf(privmsg, "PRIVMSG %s :%s", bufs[b2].bname, args);
-						irc_tx(bufs[b2].handle, privmsg);
-						while(args[strlen(args)-1]=='\n')
-							args[strlen(args)-1]=0; // stomp out trailing newlines, they break things
-						bool al=bufs[b2].alert; // save alert status...
-						int hi=bufs[b2].hi_alert;
-						add_to_buffer(b2, MSG, NORMAL, 0, true, args, SERVER(b2).nick);
-						bufs[b2].alert=al; // and restore it
-						bufs[b2].hi_alert=hi;
-					}
-					else
-					{
-						add_to_buffer(b2, ERR, NORMAL, 0, false, "Tab not live, can't send", "/amsg: ");
-					}
-				}
-			}
-		}
-		else
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify a message!", "/amsg: ");
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "me")==0)
-	{
-		if(!((bufs[cbuf].type==CHANNEL)||(bufs[cbuf].type==PRIVATE)))
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't talk here, not a channel/private chat", "/me: ");
-		}
-		else if(args)
-		{
-			if(SERVER(cbuf).handle)
-			{
-				if(LIVE(cbuf))
-				{
-					char privmsg[32+strlen(bufs[cbuf].bname)+strlen(args)];
-					sprintf(privmsg, "PRIVMSG %s :\001ACTION %s\001", bufs[cbuf].bname, args);
-					irc_tx(SERVER(cbuf).handle, privmsg);
-					while(args[strlen(args)-1]=='\n')
-						args[strlen(args)-1]=0; // stomp out trailing newlines, they break things
-					add_to_buffer(cbuf, ACT, NORMAL, 0, true, args, SERVER(cbuf).nick);
-				}
-				else
-				{
-					add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/me: ");
-				}
-			}
-			else
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't send to channel - not connected!", "/msg: ");
-			}
-		}
-		else
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify an action!", "/me: ");
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "tab")==0)
-	{
-		if(!args)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify a tab!", "/tab: ");
-		}
-		else
-		{
-			int bufn;
-			if(sscanf(args, "%d", &bufn)==1)
-			{
-				if((bufn>=0) && (bufn<nbufs))
-				{
-					cbuf=bufn;
-					redraw_buffer();
-				}
-				else
-				{
-					add_to_buffer(cbuf, ERR, NORMAL, 0, false, "No such tab!", "/tab: ");
-				}
-			}
-			else
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must specify a tab!", "/tab: ");
-			}
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "sort")==0)
-	{
-		int newbufs[nbufs];
-		buffer bi[nbufs];
-		newbufs[0]=0;
-		int b,buf=1;
-		for(b=0;b<nbufs;b++)
-		{
-			bi[b]=bufs[b];
-			if(bufs[b].type==SERVER)
-			{
-				newbufs[buf++]=b;
-				int b2;
-				for(b2=1;b2<nbufs;b2++)
-				{
-					if((bufs[b2].server==b)&&(b2!=b))
-					{
-						newbufs[buf++]=b2;
-					}
-				}
-			}
-		}
-		if(buf!=nbufs)
-		{
-			add_to_buffer(cbuf, ERR, QUIET, 0, false, "Internal error (bad count)", "/sort: ");
-			return(0);
-		}
-		int serv=0, cb=0;
-		for(b=0;b<nbufs;b++)
-		{
-			bufs[b]=bi[newbufs[b]];
-			if(bufs[b].type==SERVER)
-				serv=b;
-			bufs[b].server=serv;
-			if(newbufs[b]==cbuf)
-				cb=b;
-		}
-		cbuf=cb;
-		redraw_buffer();
-		return(0);
-	}
-	if(strcmp(cmd, "left")==0)
-	{
-		if(cbuf<2)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't move (status) tab!", "/left: ");
-		}
-		else
-		{
-			buffer tmp=bufs[cbuf];
-			bufs[cbuf]=bufs[cbuf-1];
-			bufs[cbuf-1]=tmp;
-			cbuf--;
-			int i;
-			for(i=0;i<nbufs;i++)
-			{
-				if(bufs[i].server==cbuf)
-				{
-					bufs[i].server++;
-				}
-				else if(bufs[i].server==cbuf+1)
-				{
-					bufs[i].server--;
-				}
-			}
-			redraw_buffer();
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "right")==0)
-	{
-		if(!cbuf)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Can't move (status) tab!", "/right: ");
-		}
-		else if(cbuf==nbufs-1)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Nowhere to move to!", "/right: ");
-		}
-		else
-		{
-			buffer tmp=bufs[cbuf];
-			bufs[cbuf]=bufs[cbuf+1];
-			bufs[cbuf+1]=tmp;
-			cbuf++;
-			int i;
-			for(i=0;i<nbufs;i++)
-			{
-				if(bufs[i].server==cbuf-1)
-				{
-					bufs[i].server++;
-				}
-				else if(bufs[i].server==cbuf)
-				{
-					bufs[i].server--;
-				}
-			}
-			redraw_buffer();
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "ignore")==0)
-	{
-		if(!args)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Missing arguments!", "/ignore: ");
-		}
-		else
-		{
-			char *arg=strtok(args, " ");
-			bool regex=false;
-			bool icase=false;
-			bool pms=false;
-			bool del=false;
-			while(arg)
-			{
-				if(*arg=='-')
-				{
-					if(strcmp(arg, "-i")==0)
-					{
-						icase=true;
-					}
-					else if(strcmp(arg, "-r")==0)
-					{
-						regex=true;
-					}
-					else if(strcmp(arg, "-d")==0)
-					{
-						del=true;
-					}
-					else if(strcmp(arg, "-p")==0)
-					{
-						pms=true;
-					}
-					else if(strcmp(arg, "-l")==0)
-					{
-						i_list();
-						break;
-					}
-					else if(strcmp(arg, "--")==0)
-					{
-						arg=strtok(NULL, "");
-						continue;
-					}
-				}
-				else
-				{
-					if(del)
-					{
-						if(i_cull(&bufs[cbuf].ilist, arg))
-						{
-							add_to_buffer(cbuf, STA, QUIET, 0, false, "Entries deleted", "/ignore -d: ");
-						}
-						else
-						{
-							add_to_buffer(cbuf, ERR, NORMAL, 0, false, "No entries deleted", "/ignore -d: ");
-						}
-					}
-					else if(regex)
-					{
-						name *new=n_add(&bufs[cbuf].ilist, arg, bufs[cbuf].casemapping);
-						if(new)
-						{
-							add_to_buffer(cbuf, STA, QUIET, 0, false, "Entry added", "/ignore: ");
-							new->icase=icase;
-							new->pms=pms;
-						}
-					}
-					else
-					{
-						char *isrc,*iusr,*ihst;
-						prefix_split(arg, &isrc, &iusr, &ihst);
-						if((!isrc) || (*isrc==0) || (*isrc=='*'))
-							isrc="[^!@]*";
-						if((!iusr) || (*iusr==0) || (*iusr=='*'))
-							iusr="[^!@]*";
-						if((!ihst) || (*ihst==0) || (*ihst=='*'))
-							ihst="[^@]*";
-						char expr[16+strlen(isrc)+strlen(iusr)+strlen(ihst)];
-						sprintf(expr, "^%s[_~]*!%s@%s$", isrc, iusr, ihst);
-						name *new=n_add(&bufs[cbuf].ilist, expr, bufs[cbuf].casemapping);
-						if(new)
-						{
-							add_to_buffer(cbuf, STA, QUIET, 0, false, "Entry added", "/ignore: ");
-							new->icase=icase;
-							new->pms=pms;
-						}
-					}
-					break;
-				}
-				arg=strtok(NULL, " ");
-			}
-		}
-		return(0);
-	}
-	if(strcmp(cmd, "mode")==0)
-	{
-		if(!SERVER(cbuf).handle)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must be run in the context of a server!", "/mode: ");
-		}
-		else
-		{
-			// /mode [{<user>|<banmask>|<limit>} [{\+|-}[[:alpha:]]+]]
-			if(args)
-			{
-				char *user=strtok(args, " ");
-				char *modes=strtok(NULL, "");
-				if(modes)
-				{
-					if(LIVE(cbuf))
-					{
-						if(bufs[cbuf].type==CHANNEL)
-						{
-							char mmsg[8+strlen(bufs[cbuf].bname)+strlen(modes)+strlen(user)];
-							sprintf(mmsg, "MODE %s %s %s", bufs[cbuf].bname, modes, user);
-							irc_tx(SERVER(cbuf).handle, mmsg);
-						}
-						else
-							add_to_buffer(cbuf, ERR, NORMAL, 0, false, "This is not a channel", "/mode: ");
-					}
-					else
-						add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/mode: ");
-				}
-				else
-				{
-					if(bufs[cbuf].type==CHANNEL)
-					{
-						name *curr=bufs[cbuf].nlist;
-						while(curr)
-						{
-							if(irc_strcasecmp(curr->data, user, SERVER(cbuf).casemapping)==0)
-							{
-								if(curr==bufs[cbuf].us) goto youmode;
-								char mm[12+strlen(curr->data)+curr->npfx];
-								int mpos=0;
-								sprintf(mm, "%s has mode %n-", curr->data, &mpos);
-								if(mpos)
-								{
-									for(unsigned int i=0;i<curr->npfx;i++)
-										mm[mpos++]=curr->prefixes[i].letter;
-									if(curr->npfx) mm[mpos]=0;
-									add_to_buffer(cbuf, MODE, NORMAL, 0, false, mm, "/mode: ");
-								}
-								else
-									add_to_buffer(cbuf, ERR, NORMAL, 0, false, "\"Impossible\" error (mpos==0)", "/mode: ");
-								break;
-							}
-							curr=curr->next;
-						}
-						if(!curr)
-						{
-							char mm[16+strlen(user)];
-							sprintf(mm, "No such nick: %s", user);
-							add_to_buffer(cbuf, ERR, NORMAL, 0, false, mm, "/mode: ");
-						}
-					}
-					else
-						add_to_buffer(cbuf, ERR, NORMAL, 0, false, "This is not a channel", "/mode: ");
-				}
-			}
-			else
-			{
-				youmode:
-				if(bufs[cbuf].type==CHANNEL)
-				{
-					if(bufs[cbuf].us)
-					{
-						char mm[20+strlen(SERVER(cbuf).nick)+bufs[cbuf].us->npfx];
-						int mpos=0;
-						sprintf(mm, "You (%s) have mode %n-", SERVER(cbuf).nick, &mpos);
-						if(mpos)
-						{
-							for(unsigned int i=0;i<bufs[cbuf].us->npfx;i++)
-								mm[mpos++]=bufs[cbuf].us->prefixes[i].letter;
-							if(bufs[cbuf].us->npfx) mm[mpos]=0;
-							add_to_buffer(cbuf, MODE, NORMAL, 0, true, mm, "/mode: ");
-						}
-						else
-							add_to_buffer(cbuf, ERR, NORMAL, 0, false, "\"Impossible\" error (mpos==0)", "/mode: ");
-					}
-					else
-						add_to_buffer(cbuf, ERR, NORMAL, 0, false, "\"Impossible\" error (us==NULL)", "/mode: ");
-				}
-				else
-					add_to_buffer(cbuf, ERR, NORMAL, 0, false, "This is not a channel", "/mode: ");
-			}
-		}
-		return(0);
-	}
-	if((strcmp(cmd, "cmd")==0)||(strcmp(cmd, "quote")==0))
-	{
-		if(!SERVER(cbuf).handle)
-		{
-			add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Must be run in the context of a server!", "/cmd: ");
-		}
-		else
-		{
-			bool force=false;
-			if(strncmp(args, "-f", 2)==0)
-			{
-				force=true;
-				args++;
-				while (*++args==' ');
-			}
-			if(force||LIVE(cbuf))
-			{
-				irc_tx(SERVER(cbuf).handle, args);
-				add_to_buffer(cbuf, STA, NORMAL, 0, false, args, "/cmd: ");
-			}
-			else
-			{
-				add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Tab not live, can't send", "/cmd: ");
-			}
-		}
-		return(0);
-	}
-	if(!cmd) cmd="";
-	char dstr[8+strlen(cmd)];
-	sprintf(dstr, "/%s: ", cmd);
-	add_to_buffer(cbuf, ERR, NORMAL, 0, false, "Unrecognised command!", dstr);
-	return(0);
+	return call_cmd(cmd,args,qmsg,master, fdmax);
 }
 
 void initibuf(ibuffer *i)
